backend/pyproject.toml
[tool.poetry]
name = "flask-blog"
version = "0.1.0"
description = ""
authors = ["Peskov Sergey <devpeskov@proton.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.2"
flask-sqlalchemy = "^3.0.2"
psycopg2 = "^2.9.5"
flask-migrate = "^4.0.0"
wtforms = "^3.0.1"
python-dotenv = "^0.21.0"
flask-admin = "^1.6.0"


[tool.poetry.group.dev.dependencies]
pynvim = "^0.4.3"
isort = "^5.11.4"
black = "^22.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
